Add basic transaction support via sequelize.startTransaction.
It uses a mechanism similar to the QueryChainer - mulitple functions are
added to the transaction and if they all succeed then the transaction is
commited and if one fails then the transaction is rolled-back and the
remaining actions (if any) are skipped.

If an update or insert happens during a transaction and the transaction
is then reverted, the model objects for the inserted/updated rows are
left in an invalid state (e.g. they are not reverted too).